Update paths for libraries, add NX [TEA-83]
