+++ .prettierignore
@@ -1,0 +1,3 @@
+# Add files here to ignore them from prettier formatting
+/dist
+/coverage
