+++ package.json
@@ -1,24 +1,17 @@
 {
-  "name": "client-server-template-with-vite",
+  "name": "workspace",
   "private": true,
   "scripts": {
     "bootstrap": "yarn && node init.js && lerna clean && yarn && lerna bootstrap",
-
-    "build:shared": "lerna run build --scope=@workspace/shared",
     "build:server": "lerna run build --scope=server",
     "build:client": "lerna run build --scope=client",
     "build": "lerna run build",
-
     "dev:client": "lerna run dev --scope=client",
     "dev:server": "lerna run dev --scope=server",
     "dev": "lerna run dev",
-
     "test": "lerna run test",
-
     "lint": "lerna run lint",
-    
     "format": "lerna run format",
-
     "preview": "lerna run preview"
   },
   "license": "MIT",
@@ -30,6 +23,21 @@
   },
   "devDependencies": {
     "@evilmartians/lefthook": "^1.3.9",
-    "lerna": "^5.4.3"
+    "@nx/jest": "16.5.5",
+    "@nx/js": "^16.5.5",
+    "@types/jest": "^29.4.0",
+    "@types/node": "18.7.1",
+    "jest": "^29.4.1",
+    "jest-environment-node": "^29.4.1",
+    "lerna": "^5.4.3",
+    "nx": "16.5.5",
+    "nx-cloud": "16.1.1",
+    "prettier": "^2.6.2",
+    "ts-jest": "^29.1.0",
+    "ts-node": "10.9.1",
+    "typescript": "~5.1.3"
+  },
+  "dependencies": {
+    "tslib": "^2.3.0"
   }
 }
