+++ packages/client/package.json
@@ -51,8 +51,7 @@
     "typescript": "^4.8.2",
     "vite": "^3.1.0",
     "vite-plugin-pwa": "^0.16.4",
-    "vite-plugin-svgr": "^3.2.0",
-    "@workspace/shared": "^1.0.0"
+    "vite-plugin-svgr": "^3.2.0"
   },
   "license": "MIT"
 }
