+++ packages/client/src/api/ViewerAPI.ts
@@ -1,5 +1,5 @@
 import { BaseApi } from '@api/BaseAPI';
-import { API_URL } from '@shared/config/constants';
+import { API_URL } from '@workspace/shared';
 
 export interface GetViewerResponse {
   id: number;
