+++ packages/client/src/api/utils.ts
@@ -1,4 +1,4 @@
-import { RESOURCES_URL } from '@shared/config/constants';
+import { RESOURCES_URL } from '@workspace/shared';
 
 export const getResourceUrl = (resource: string) => {
   return `${RESOURCES_URL}/${resource}`;
