+++ packages/client/src/controllers/AuthController.ts
@@ -1,6 +1,6 @@
 import { ISignupData, ISigninData } from '@api/AuthAPI';
 import { GetViewerResponse } from '@api/ViewerAPI';
-import { setUser } from '@shared/store/reducers/userSlice';
+import { setUser } from '@workspace/shared';
 
 import { Handlers, BaseController } from './BaseController';
