+++ packages/client/src/controllers/BaseController.ts
@@ -1,5 +1,5 @@
 import { ServicesModel } from '@core/ServicesContext';
-import { RootState, RootStore } from '@shared/store';
+import { RootState, RootStore } from '@workspace/shared';
 
 export interface Handlers<T = unknown> {
   onSuccess?: (data: T) => void;
