--- packages/server/index.ts
@@ -1,7 +1,0 @@
-import * as dotenv from 'dotenv';
-
-import { createApp } from '@app';
-
-dotenv.config();
-
-createApp();
