+++ packages/server/package.json
@@ -5,7 +5,7 @@
   "scripts": {
     "build": "rimraf dist && tsc --p ./tsconfig.prod.json",
     "preview": "node ./dist/server/index.js",
-    "dev": "cross-env NODE_ENV=development nodemon index.ts",
+    "dev": "cross-env NODE_ENV=development nodemon src/index.ts",
     "lint": "eslint .",
     "format": "prettier --write .",
     "test": "jest ."
@@ -32,13 +32,12 @@
     "@types/cors": "^2.8.12",
     "@types/express": "^4.17.13",
     "@types/jest": "^28.1.8",
-    "@types/node-fetch": "^2.6.4",
     "@types/node": "^20.4.2",
+    "@types/node-fetch": "^2.6.4",
     "@types/pg": "^8.6.5",
     "@types/request": "^2.48.8",
     "@typescript-eslint/eslint-plugin": "^5.35.1",
     "@typescript-eslint/parser": "^5.35.1",
-    "@workspace/shared": "^1.0.0",
     "babel-jest": "^29.0.1",
     "eslint": "^8.23.0",
     "eslint-config-prettier": "^8.5.0",
