--- packages/server/src/@types/index.d.ts
@@ -1,4 +1,0 @@
-import '@workspace/shared/@types';
-
-export * from './env';
-export * from './express-extends';
