+++ packages/server/src/index.ts
@@ -1,0 +1,7 @@
+import * as dotenv from 'dotenv';
+
+import { createApp } from '@app';
+
+dotenv.config();
+
+createApp();
