+++ packages/server/tsconfig.prod.json
@@ -1,4 +1,4 @@
 {
-  "extends": "./tsconfig",
+  "extends": "./tsconfig.json",
   "exclude": ["**/*.test.ts", "**/*.mock.ts", "**/__test__"]
 }
