--- packages/shared/@types/index.d.ts
@@ -1,1 +1,0 @@
-import './typed';
