--- packages/shared/@types/typed.d.ts
@@ -1,7 +1,0 @@
-import { RootState } from '../store';
-
-declare global {
-  interface Window {
-    __REDUX_STORE__?: RootState;
-  }
-}
