--- packages/shared/config/constants.ts
@@ -1,2 +1,0 @@
-export const API_URL = 'https://ya-praktikum.tech/api/v2';
-export const RESOURCES_URL = 'https://ya-praktikum.tech/api/v2/resources';
