--- packages/shared/models/viewer.ts
@@ -1,10 +1,0 @@
-export interface Viewer {
-  id: number;
-  email: string | null;
-  login: string | null;
-  first_name: string | null;
-  second_name: string | null;
-  display_name: string | null;
-  phone: string | null;
-  avatar: string | null;
-}
