+++ packages/shared/package.json
@@ -1,6 +1,5 @@
 {
   "name": "@workspace/shared",
   "version": "1.0.0",
-  "main": "index.js",
-  "license": "MIT"
+  "main": "src/index.js"
 }
