+++ packages/shared/src/@types/index.d.ts
@@ -1,0 +1,7 @@
+import { RootState } from '../store';
+
+declare global {
+  interface Window {
+    __REDUX_STORE__?: RootState;
+  }
+}
