+++ packages/shared/src/config/constants.ts
@@ -1,0 +1,2 @@
+export const API_URL = 'https://ya-praktikum.tech/api/v2';
+export const RESOURCES_URL = 'https://ya-praktikum.tech/api/v2/resources';
