+++ packages/shared/src/store/reducers/index.ts
@@ -1,0 +1,1 @@
+export * from './userSlice';
