--- packages/shared/store/hooks/index.ts
@@ -1,6 +1,0 @@
-import { TypedUseSelectorHook, useDispatch, useSelector } from 'react-redux';
-
-import { RootState, AppDispatch } from '..';
-
-export const useAppDispatch: () => AppDispatch = useDispatch;
-export const useAppSelector: TypedUseSelectorHook<RootState> = useSelector;
